refactoring widgets
