+++ lib/screens/profile/set_profile_screen.dart
@@ -2,10 +2,10 @@ import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:partypal/configs/router_config.dart';
 import 'package:partypal/constants/route_paths.dart';
-import 'package:partypal/widgets/app_bar.dart';
-import 'package:partypal/widgets/buttons.dart';
-import 'package:partypal/widgets/edit_picture.dart';
-import 'package:partypal/widgets/tonal_elevation.dart';
+import 'package:partypal/widgets/app_bars/app_bar.dart';
+import 'package:partypal/widgets/buttons/wide_button.dart';
+import 'package:partypal/widgets/sheets/edit_picture.dart';
+import 'package:partypal/widgets/others/tonal_elevation.dart';
 
 class SetProfileScreen extends StatefulWidget {
   const SetProfileScreen({super.key});
