+++ lib/widgets/app_bars/app_bar.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:partypal/configs/router_config.dart';
 import 'package:partypal/constants/asset_paths.dart';
-import 'package:partypal/widgets/tonal_elevation.dart';
+import 'package:partypal/widgets/others/tonal_elevation.dart';
 
 class CustomAppBar extends StatelessWidget {
   final String title;
